# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: translator-server API
  version: 1.0.0
  summary: Canonical REST API for the Yara translator-server.
  description: |
    The canonical API surface is mounted under `/api/v1`. Legacy aliases under
    `/api/db/*`, `/api/user/*`, `/api/epubs/*`, `/api/backup/*`,
    `/api/browser-workers`, `/api/proxy/*`, `/api/defaults`,
    `/api/translation-jobs/*` are **deprecated** and will be removed on
    **Wed, 01 Jan 2026** (the `Sunset` header is set on every legacy response).
    New clients must use `/api/v1`.

    When the running server and this document diverge, the running server is the
    source of truth.
  contact:
    name: Yara
    url: https://github.com/mfloresz/yara
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api/v1
    description: Canonical v1 surface
  - url: /
    description: Server root (legacy aliases live directly under /api/...)
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /novels:
    get:
      tags: [Novels]
      summary: List the user's novels.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/FieldsParam'
        - $ref: '#/components/parameters/SelectParam'
        - name: q
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string }
        - name: order
          in: query
          schema: { type: string, enum: [asc, desc] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CollectionEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Novel' }
    post:
      tags: [Novels]
      summary: Create a novel.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NovelCreate' }
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema: { type: string }
              description: Canonical URL of the new resource, e.g. `/api/v1/novels/<id>`.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Novel' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /novels/tags/suggestions:
    get:
      tags: [Novels]
      summary: Distinct tag values for autocomplete.
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CollectionEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { type: string }
  /novels/series/suggestions:
    get:
      tags: [Novels]
      summary: Distinct series values for autocomplete.
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CollectionEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { type: string }
  /novels/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Novels]
      summary: Get one novel.
      parameters:
        - $ref: '#/components/parameters/FieldsParam'
        - $ref: '#/components/parameters/SelectParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Novel' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Novels]
      summary: Partial update.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NovelPatch' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Novel' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Novels]
      summary: Delete the novel and all its chapters.
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '404': { $ref: '#/components/responses/NotFound' }
  /novels/{id}/clone:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      tags: [Novels]
      summary: Duplicate a novel.
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Novel' }
  /novels/{id}/visibility:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    patch:
      tags: [Novels]
      summary: Toggle public visibility.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isPublic]
              properties:
                isPublic: { type: boolean }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Novel' }
  /novels/{id}/cover:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      tags: [Novels]
      summary: Upload a cover image.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [cover]
              properties:
                cover:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Novel' }
  /novels/{id}/recalculate-stats:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      tags: [Novels]
      summary: Recompute chapter counts and char counts.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Novel' }
  /novels/{id}/full:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Novels]
      summary: Return the novel plus all its chapters (heavy).
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          novel: { $ref: '#/components/schemas/Novel' }
                          chapters:
                            type: array
                            items: { $ref: '#/components/schemas/Chapter' }

  # ---------------------------------------------------------------------------
  # Chapters
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth.token
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string }
...
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PerPageParam:
      name: per_page
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200 }
    OffsetParam:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    FieldsParam:
      name: fields
      in: query
      description: Comma-separated list of allowed fields. Heavy fields (coverPath, glossary, tags, aiOptions, translationOptions, cleanupRules, sourceDescription, targetDescription, notes) are excluded unless requested.
      schema: { type: string }
    SelectParam:
      name: select
      in: query
      description: Compat alias for `fields`.
      schema: { type: string }
  responses:
    NoContent:
      description: No Content.
    BadRequest:
      description: Bad request.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    NotFound:
      description: Not found.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      description: RFC 7807 problem details. The envelope uses `error` instead of the standard top-level fields to keep the v1 envelope consistent.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [bad_request, unauthorized, forbidden, not_found, conflict, queue_full, internal_error]
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
                  code: { type: string }
    SingleEnvelope:
      type: object
      description: Canonical v1 single-resource envelope.
      required: [data]
      properties:
        data: {}
    CollectionEnvelope:
      type: object
      description: Canonical v1 collection envelope.
      required: [data, meta, links]
      properties:
        data:
          type: array
          items: {}
        meta:
          $ref: '#/components/schemas/Meta'
        links:
          $ref: '#/components/schemas/Links'
    Meta:
      type: object
      required: [total, has_more]
      properties:
        total: { type: integer, format: int32 }
        page: { type: integer, format: int32 }
        per_page: { type: integer, format: int32 }
        limit: { type: integer, format: int32 }
        offset: { type: integer, format: int32 }
        has_more: { type: boolean }
        next_page: { type: [integer, 'null'], format: int32 }
    Links:
      type: object
      required: [self]
      properties:
        self: { type: string }
        next: { type: string }
        prev: { type: string }
        first: { type: string }
        last: { type: string }
...
    Novel:
      type: object
      description: |
        The full novel record. The list endpoint with `?fields=` returns a subset
        of these properties. Heavy fields: coverPath, glossary, tags, aiOptions,
        translationOptions, cleanupRules, sourceDescription, targetDescription,
        notes.
      required: [id, sourceTitle, status, sourceLanguage, targetLanguage]
      properties:
        id: { type: string }
        ownerId: { type: string }
        sourceLanguage: { type: string }
        targetLanguage: { type: string }
        sourceTitle: { type: string }
        sourceAuthor: { type: string }
        sourceDescription: { type: string }
        sourceSeries: { type: string }
        sourceNumber: { type: string }
        targetTitle: { type: string }
        targetAuthor: { type: string }
        targetDescription: { type: string }
        targetSeries: { type: string }
        targetNumber: { type: string }
        url: { type: string, format: uri }
        customCommands: { type: string }
        status:
          type: string
          enum: [ongoing, completed, hiatus, dropped]
        coverPath: { type: string, description: Thumbnail path; falls back to the full cover. }
        isPublic: { type: boolean }
        chapterCount: { type: integer }
        translatedCount: { type: integer }
        completedCount: { type: integer }
        originalCharCount: { type: integer }
        translatedCharCount: { type: integer }
        refinedCharCount: { type: integer }
        totalCharCount: { type: integer }
        maxChapterOrder: { type: integer }
        lastCheckedAt: { type: [string, 'null'], format: date-time }
        lastCheckNewChapters: { type: integer }
        lastReadAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string }
        updatedAt: { type: string }
        glossary:
          type: array
          items: { type: object }
        tags:
          type: array
          items: { type: string }
        prompts: { type: object }
        aiOptions: { type: object }
        translationOptions: { type: object }
        cleanupRules:
          type: array
          items: { type: object }
        notes: { type: string }
        canUpdate:
          type: boolean
          description: Whether the URL is in the parser catalog. Only set when explicitly requested via `?fields=canUpdate`.
        requiresBrowser:
          type: boolean
          description: Whether the URL needs the browser-worker extension. Only set when explicitly requested via `?fields=requiresBrowser`.
    NovelCreate:
      type: object
      required: [sourceTitle]
      properties:
        sourceLanguage: { type: string }
        targetLanguage: { type: string }
        sourceTitle: { type: string }
        sourceAuthor: { type: string }
        sourceDescription: { type: string }
        sourceSeries: { type: string }
        sourceNumber: { type: string }
        targetTitle: { type: string }
        targetAuthor: { type: string }
        targetDescription: { type: string }
        targetSeries: { type: string }
        targetNumber: { type: string }
        glossary: { type: array, items: { type: object } }
        prompts: { type: object }
        notes: { type: string }
        aiOptions: { type: object }
        translationOptions: { type: object }
        cleanupRules: { type: array, items: { type: object } }
        url: { type: string, format: uri }
        customCommands: { type: string }
        tags:
          type: array
          items: { type: string }
    NovelPatch:
      type: object
      description: Any subset of Novel's writable fields.
      additionalProperties: true
    Chapter:
      type: object
      required: [id, novelId, chapterOrder]
      properties:
        id: { type: string }
        novelId: { type: string }
        chapterOrder: { type: integer }
        title: { type: string }
        translatedTitle: { type: string }
        originalContent: { type: string }
        translatedContent: { type: string }
        refinedContent: { type: string }
        status:
          type: string
          enum: [pending, processing, completed, failed, skipped]
        errorMessage: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
...